test: remove unreachable factory functional helpers
