+++ docs/internal/baselines/deadcode-baseline.txt
@@ -160,21 +160,6 @@ tests/functional/bootstrap_portability/automat_portability_fixture_test.go: unre
 tests/functional/bootstrap_portability/export_import_fixture_test.go: unreachable func: exportImportFixture.persistAs
 tests/functional/cli/mcp_resume/serve_runtime_resume_smoke_test.go: unreachable func: mcpControlResume
 tests/functional/cli/mcp_resume/stdio_client_test.go: unreachable func: assertInstallSmokeDiscovery
-tests/functional/factory/current/helpers_test.go: unreachable func: activateNamedPersistedFactoryOverHTTP
-tests/functional/factory/current/helpers_test.go: unreachable func: assertCurrentFactoryNameAndDirectory
-tests/functional/factory/current/helpers_test.go: unreachable func: assertNoAdditionalProviderCommandWork
-tests/functional/factory/current/helpers_test.go: unreachable func: factoryHasCustomerPlace
-tests/functional/factory/current/helpers_test.go: unreachable func: functionalNamedFactoryPayloadWithTerminalState
-tests/functional/factory/current/helpers_test.go: unreachable func: namedFilewatcherFactoryPayloadWithProcess
-tests/functional/factory/current/helpers_test.go: unreachable func: seedFilewatcherNamedFactoryRootWithProcess
-tests/functional/factory/current/helpers_test.go: unreachable func: seedNamedFactoryRootWithTerminalStateAndProcess
-tests/functional/factory/current/helpers_test.go: unreachable func: startCurrentFactoryServerWithProviderRunnerAndSetup
-tests/functional/factory/current/helpers_test.go: unreachable func: waitForActivatedFactoryRuntime
-tests/functional/factory/current/helpers_test.go: unreachable func: waitForProviderCommandWorkSettlement
-tests/functional/factory/current/helpers_test.go: unreachable func: waitForSubmittedWorkAtPlace
-tests/functional/factory/current/helpers_test.go: unreachable func: workCustomerPlaceID
-tests/functional/factory/current/helpers_test.go: unreachable func: workTraceID
-tests/functional/factory/packaged/subagent/invocation_test.go: unreachable func: packagedSubagentTextInvocationRequest
 tests/functional/internal/support/api_server.go: unreachable func: FunctionalAPIServer.WaitForExitError
 tests/functional/internal/support/api_server.go: unreachable func: StartFunctionalAPIServiceModeServer
 tests/functional/internal/support/events.go: unreachable func: LastFactoryEventTick
